refactor: move lnurl-channel handling to lnurl reducer
