+++ renderer/components/Channels/index.js
@@ -33,4 +33,3 @@ export ChannelsSummaryDonut from './ChannelsSummaryDonut'
 export ChannelsViewSwitcher from './ChannelsViewSwitcher'
 export NodeCardList from './NodeCardList'
 export NodeCardView from './NodeCardView'
-export LnurlChannelPrompt from './LnurlChannelPrompt'
