+++ renderer/components/Lnurl/index.js
@@ -1,1 +1,2 @@
 export LnurlAuthPrompt from './LnurlAuthPrompt'
+export LnurlChannelPrompt from './LnurlChannelPrompt'
