+++ renderer/components/Lnurl/messages.js
@@ -6,4 +6,8 @@ export default defineMessages({
   lnurl_auth_prompt_dialog_body: 'Do you confirm authentication request from {service}',
   lnurl_auth_prompt_dialog_confirm_text: 'Confirm',
   lnurl_auth_prompt_dialog_decline_text: 'Decline',
+  lnurl_channel_prompt_dialog_header: 'Confirm channel',
+  lnurl_channel_prompt_dialog_body: 'Do you confirm channel request to {service}',
+  lnurl_channel_prompt_dialog_confirm_text: 'Confirm',
+  lnurl_channel_prompt_dialog_decline_text: 'Decline',
 })
