+++ renderer/reducers/channels/constants.js
@@ -28,6 +28,3 @@ export const CHANNELS_SORT_LOCAL_BALANCE = 'CHANNELS_SORT_LOCAL_BALANCE'
 export const CHANNELS_SORT_ACTIVITY = 'CHANNELS_SORT_ACTIVITY'
 export const CHANNELS_SORT_NAME = 'CHANNELS_SORT_NAME'
 export const CHANNELS_SORT_CAPACITY = 'CHANNELS_SORT_CAPACITY'
-
-export const SET_LNURL_CHANNEL_PARAMS = 'SET_LNURL_CHANNEL_PARAMS'
-export const CLEAR_LNURL_CHANNEL = 'CLEAR_LNURL_CHANNEL'
