+++ renderer/reducers/channels/index.js
@@ -4,5 +4,4 @@ export default channelsReducer
 export channelsSelectors from './selectors'
 export * from './constants'
 export * from './reducer'
-export * from './ipc'
 export * from './utils'
