+++ renderer/reducers/channels/selectors.js
@@ -34,7 +34,6 @@ const channelSearchQuerySelector = state => state.channels.searchQuery
 const channelSortSelector = state => state.channels.sort
 const channelSortOrderSelector = state => state.channels.sortOrder
 const filterSelector = state => state.channels.filter
-const lnurlChannelParams = state => state.channels.lnurlChannelParams
 const nodesSelector = state => networkSelectors.nodes(state)
 
 export const channelMatchesQuery = (channelObj, searchQuery) => {
@@ -373,17 +372,5 @@ channelsSelectors.isCustomFilter = createSelector(filterSelector, filters => {
   const difference = new Set([...filters].filter(x => !DEFAULT_FILTER.has(x)))
   return difference.size > 0
 })
-
-/**
- * willShowLnurlChannelPrompt - Boolean indicating wether lnurl channel prompt should show.
- */
-channelsSelectors.willShowLnurlChannelPrompt = createSelector(lnurlChannelParams, params => {
-  return Boolean(params)
-})
-
-/**
- * lnurlChannelParams - Current lnurl withdrawal paramaters.
- */
-channelsSelectors.lnurlChannelParams = createSelector(lnurlChannelParams, params => params)
 
 export default channelsSelectors
