+++ renderer/reducers/lnurl/constants.js
@@ -1,2 +1,5 @@
 export const SET_LNURL_AUTH_PARAMS = 'SET_LNURL_AUTH_PARAMS'
 export const CLEAR_LNURL_AUTH = 'CLEAR_LNURL_AUTH'
+
+export const SET_LNURL_CHANNEL_PARAMS = 'SET_LNURL_CHANNEL_PARAMS'
+export const CLEAR_LNURL_CHANNEL = 'CLEAR_LNURL_CHANNEL'
