+++ renderer/reducers/lnurl/reducer.js
@@ -2,11 +2,20 @@ import { send } from 'redux-electron-ipc'
 import CryptoJS from 'crypto-js'
 import createReducer from '@zap/utils/createReducer'
 import { wordArrayToByteArray, byteToHexString } from '@zap/utils/byteutils'
+import { getIntl } from '@zap/i18n'
 import { grpc } from 'workers'
+import { infoSelectors } from 'reducers/info'
+import { showWarning } from 'reducers/notification'
 import * as constants from './constants'
 import lnurlSelectors from './selectors'
+import messages from './messages'
 
-const { SET_LNURL_AUTH_PARAMS, CLEAR_LNURL_AUTH } = constants
+const {
+  SET_LNURL_AUTH_PARAMS,
+  CLEAR_LNURL_AUTH,
+  SET_LNURL_CHANNEL_PARAMS,
+  CLEAR_LNURL_CHANNEL,
+} = constants
 
 // ------------------------------------
 // Initial State
@@ -15,15 +24,17 @@ const { SET_LNURL_AUTH_PARAMS, CLEAR_LNURL_AUTH } = constants
 /**
  * @typedef State
  * @property {{service: string, secret: string}|null} lnurlAuthParams Current lnurl auth paramaters.
+ * @property {{service: string, secret: string}|null} lnurlChannelParams Current lnurl channel paramaters.
  */
 
 /** @type {State} */
 const initialState = {
   lnurlAuthParams: null,
+  lnurlChannelParams: null,
 }
 
 // ------------------------------------
-// Actions
+// Actions - lnurl-auth
 // ------------------------------------
 
 /**
@@ -110,6 +121,69 @@ export const finishLnurlAuth = () => async (dispatch, getState) => {
 }
 
 // ------------------------------------
+// Actions - lnurl-channel
+// ------------------------------------
+
+/**
+ * setLnurlChannelParams - Set request details.
+ *
+ * @param {object|null} params lnurl request details or null to clear
+ * @returns {object} Action
+ */
+export function setLnurlChannelParams(params) {
+  return {
+    type: SET_LNURL_CHANNEL_PARAMS,
+    params,
+  }
+}
+
+/**
+ * clearLnurlChannel - Clears lnurl channel state
+ *
+ * @returns {(dispatch:Function) => void} Thunk
+ */
+export const clearLnurlChannel = () => dispatch => {
+  dispatch({ type: CLEAR_LNURL_CHANNEL })
+}
+
+/**
+ * declineLnurlChannel - Cancels lnurl channel and clears params cache.
+ *
+ * @returns {(dispatch:Function) => void} Thunk
+ */
+export const declineLnurlChannel = () => dispatch => {
+  dispatch(send('lnurlCancelChannel'))
+  dispatch(clearLnurlChannel())
+}
+
+/**
+ * finishLnurlChannel - Concludes lnurl channel request processing by sending our ln pubkey to the service.
+ *
+ * @returns {(dispatch:Function, getState:Function) => Promise<void>} Thunk
+ */
+export const finishLnurlChannel = () => async (dispatch, getState) => {
+  const state = getState()
+  const lnurlChannelParams = lnurlSelectors.lnurlChannelParams(state)
+  if (lnurlChannelParams) {
+    const { service } = lnurlChannelParams
+
+    // Show notification.
+    dispatch(
+      showWarning(getIntl().formatMessage(messages.lnurl_channel_started), {
+        timeout: 3000,
+        isProcessing: true,
+        payload: { service },
+      })
+    )
+
+    const [pubkey, host] = service.split('@')
+    await grpc.services.Lightning.ensurePeerConnected({ pubkey, host })
+    dispatch(send('lnurlFinishChannel', { pubkey: infoSelectors.nodePubkey(state) }))
+    dispatch(clearLnurlChannel())
+  }
+}
+
+// ------------------------------------
 // Action Handlers
 // ------------------------------------
 
@@ -119,6 +193,12 @@ const ACTION_HANDLERS = {
   },
   [CLEAR_LNURL_AUTH]: state => {
     state.lnurlAuthParams = null
+  },
+  [SET_LNURL_CHANNEL_PARAMS]: (state, { params }) => {
+    state.lnurlChannelParams = params
+  },
+  [CLEAR_LNURL_CHANNEL]: state => {
+    state.lnurlChannelParams = null
   },
 }
